New base structure created
MachineState is a class to implements the automa: has a doExecution method to call to execute the current state. This one is a method to implement with at last expression the call to doTransition to set the next state. The structure works with function pointers defined as ExecStateFunct type, so it's static. Every state has a method to call and a relative constant to refer to it into the program. New states has to have their method and constant.

First state is STATE_IDLE

The data of the program are into the singleton WorkingSet (static), the sensor's actions are into the singleton Trasductor (statis) and the device's actions are into the singleton Actuator (static)

Interrupt can change the next state and disable the actions on the devices